helpers: rename mcu to ec_cmd
